return error page test

+++ 3.00-starting-project/src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -34,7 +34,10 @@ public class GradebookController {
     }
 
     @GetMapping("/delete/student/{id}")
-    public String deleteStudent(@PathVariable int id, Model m){
+    public String deleteStudent(@PathVariable Integer id, Model m){
+        if( !studentAndGradeService.checkIfStudentIsNull(id)){
+            return "error";
+        }
         studentAndGradeService.deleteStudent(id);
         Iterable<CollegeStudent> collegeStudents = studentAndGradeService.getGradebook();
         m.addAttribute("students", collegeStudents);
